backend/be/requirements.txt
uvicorn
fastapi
mysqlclient
sqlalchemy
PyMySQL
python-multipart
python-jose[cryptography]
passlib[bcrypt]